Fixes #56: Changed the bean class name and added another constructor
